requirements.txt
# ============================================================
# ecom-lead-gen-system/requirements.txt
# Pin exact versions for reproducible production deployments
# ============================================================

# --- Web Scraping ---
playwright==1.42.0

# --- HTTP ---
requests==2.31.0
httpx==0.27.0

# --- AI ---
google-generativeai==0.8.3

# --- Data Validation ---
pydantic==2.6.4
pydantic-settings==2.2.1

# --- Storage ---
pyairtable==2.2.2

# --- Environment ---
python-dotenv==1.0.1

# --- Resilience & Retry ---
tenacity==8.2.3

# --- Scheduling ---
APScheduler==3.10.4

# --- Notifications ---
slack-sdk==3.27.1

# --- Console / Logging ---
rich==13.7.1

# --- Testing ---
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
responses==0.25.3
freezegun==1.4.0

# --- Type Stubs (optional, for mypy) ---
types-requests==2.31.0.20240311
